Updated the home page to warn against running prod code
Updated the home page to include a warning that this is not a
production system and that production code and data should not be run
from the platform.

How to test: run hugo in server mode
navigate to localhost/index.html
The welcome page should have text encouraging users not to deploy
production applications.