Add deprecation notice to README
When merged I will prefix the name of this repo with `paas-alpha-` and
transfer ownership to https://github.com/gds-attic

GitHub will redirect requests for the old name(s) to the new location.

NB: Unlike alphagov/cf-terraform, I'm referring to the new docs rather than
alphagov/paas-cf, because that may be what someone is specifically looking
for.